функция получения даты вынесена в отдельный модуль
